requirements.txt
Flask==2.3.2
pandas==2.1.1
openpyxl==3.1.2
python-docx==0.8.11
gunicorn==20.1.0